+++ src/main/java/janggi/view/Parser.java
@@ -1,4 +1,4 @@
-package janggi.util;
+package janggi.view;
 
 import java.util.Arrays;
 import java.util.List;
